% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addBlusters.R
\name{addBlusters}
\alias{addBlusters}
\title{flexibly add clusterings with a variety of methods and benchmarks}
\usage{
addBlusters(x, BLUSPARAM = NULL, rdName = NULL)
}
\arguments{
\item{x}{a SingleCellExperiment object or something like it}

\item{BLUSPARAM}{a BlusterParam object (defaults are often workable)}

\item{rdName}{which reducedDim to use? (the first one)}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{         a factor of length(nrow(reducedDim
}\if{html}{\out{</div>}}
}
\description{
flexibly add clusterings with a variety of methods and benchmarks
}
\details{
this function delegates to clusterRows(), which as the name
suggests, works on a matrix where the observations are rows and
the measurements are columns (i.e., a reducedDim). If no BLUSPARAM
is specified, the function defaults to
NNGraphParam(cluster.fun='louvain',
cluster.args=list(resolution=0.8))
}
\seealso{
bluster::clusterRows

bluster::clusterSweep

bluster::approxSilhouette

bluster::bootstrapStability

bluster::compareClusterings

bluster::linkClusters

bluster::NNGraphParam-class

bluster::DbscanParam-class
}
