% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addIgvTrack.R
\name{addIgvTrack}
\alias{addIgvTrack}
\title{wrapper to add a weights track to an igvR instance}
\usage{
addIgvTrack(x, trk, igv, co = "violet", a = TRUE, ...)
}
\arguments{
\item{x}{SingleCellExperiment with weights in mcols(x)$trk}

\item{trk}{name of the column in mcols(x) with weights to plot}

\item{igv}{an igvR object with an appropriate genome to match genome(x)}

\item{co}{color for the track (violet)}

\item{a}{autoscale the track? (TRUE)}

\item{...}{additional parameters passed to igvR::GRangesQuantitativeTrack}
}
\value{
nothing useful (called for side effects)
}
\description{
wrapper to add a weights track to an igvR instance
}
\examples{
library(igvR)
igv <- igvR()
setGenome(igv, "hg38")
mcols(SCE)$aTrack <- rpois(nrow(SCE), lambda=50)
addIgvTrack(SCE, "aTrack", igv) 

}
\seealso{
igvR::GRangesQuantitativeTrack

igvR::displayTrack

igvR::setGenome

igvR::igvR
}
