% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addNMF.R
\name{addNMF}
\alias{addNMF}
\title{add an NMF decomposition to the reducedDims and rowRanges of an SCE}
\usage{
addNMF(x, k = 30, colDat = TRUE, rowDat = TRUE, ...)
}
\arguments{
\item{x}{a SingleCellExperiment, usually from archRtoSCE}

\item{k}{rank for the NMF decomposition on logCounts (30)}

\item{colDat}{add NMF scores to colData(x)? (TRUE)}

\item{rowDat}{add NMF weights to mcols(rowRanges(x))? (TRUE)}

\item{...}{additional arguments to pass to RcppML::nmf()}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{         SingleCellExperiment with reducedDim(x, "NMF")
}\if{html}{\out{</div>}}
}
\description{
By default, the weights for each factor are copied to mcols(rowRanges(x)).
}
\details{
\if{html}{\out{<div class="sourceCode">}}\preformatted{        the mcols(rowRanges(x)) weights are for igvR plotting 
}\if{html}{\out{</div>}}
}
\seealso{
\if{html}{\out{<div class="sourceCode">}}\preformatted{        archRtoSCE
}\if{html}{\out{</div>}}

\if{html}{\out{<div class="sourceCode">}}\preformatted{        iSEEarchR
}\if{html}{\out{</div>}}

\if{html}{\out{<div class="sourceCode">}}\preformatted{        addIgvTrack
}\if{html}{\out{</div>}}
}
