% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addTfIdf.R
\name{addTfIdf}
\alias{addTfIdf}
\title{add ArchR-style log1p(TF-IDF) transformed counts matrix to assays(x, 'TFIDF')}
\usage{
addTfIdf(x, idf = NULL, ...)
}
\arguments{
\item{x}{a SummarizedExperiment}

\item{idf}{a pre-trained idf or logTfIdf result (NULL; compute idf)}

\item{...}{additional arguments passed to logTfIdf()}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{     x, but with a TFIDF matrix added to assays(x)
}\if{html}{\out{</div>}}
}
\description{
add ArchR-style log1p(TF-IDF) transformed counts matrix to assays(x, 'TFIDF')
}
\details{
\if{html}{\out{<div class="sourceCode">}}\preformatted{    the 'idf' parameter for logTfIdf() is of interest, as
             it will be propagated into metadata(x)$idf and can be
             passed in to addTfIdf as well (e.g. if projecting data)
}\if{html}{\out{</div>}}
}
\seealso{
\if{html}{\out{<div class="sourceCode">}}\preformatted{    logTfIdf
}\if{html}{\out{</div>}}
}
