% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addUMAP.R
\name{addUMAP}
\alias{addUMAP}
\title{add a uwot::umap projection to reducedDims(x) AND SAVE THE MODEL TO METADATA}
\usage{
addUMAP(
  x,
  reducedDims = NULL,
  name = "UMAP",
  nNeighbors = 40,
  minDist = 0.4,
  metric = "cosine",
  scale = "range",
  ...
)
}
\arguments{
\item{x}{a SingleCellExperiment, usually from archRtoSCE}

\item{reducedDims}{name of reducedDim used (default: NMF if found else LSI)}

\item{name}{name of reducedDim to store the UMAP embedding ("UMAP")}

\item{nNeighbors}{integer nearest neighbors to use for UMAP (40)}

\item{minDist}{passed to uwot::umap(min_dist); controls packing (0.4)}

\item{metric}{distance metric for uwot::umap to use ("cosine")}

\item{...}{additional arguments to pass to uwot::umap()}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{         SingleCellExperiment with reducedDim(x, "NMF")
}\if{html}{\out{</div>}}
}
\description{
add a uwot::umap projection to reducedDims(x) AND SAVE THE MODEL TO METADATA
}
\details{
\if{html}{\out{<div class="sourceCode">}}\preformatted{        uses the same default neighbors (40) as ArchR by default
}\if{html}{\out{</div>}}
}
\seealso{
\if{html}{\out{<div class="sourceCode">}}\preformatted{        addNMF
}\if{html}{\out{</div>}}

\if{html}{\out{<div class="sourceCode">}}\preformatted{        iSEEarchR
}\if{html}{\out{</div>}}

\if{html}{\out{<div class="sourceCode">}}\preformatted{        uwot::umap 
}\if{html}{\out{</div>}}
}
