% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archRbarCodes.R
\name{archRbarCodes}
\alias{archRbarCodes}
\title{extract unique barcodes from cellColData in an archR project}
\usage{
archRbarCodes(proj, splt = "#", pos = 2)
}
\arguments{
\item{proj}{the project}

\item{splt}{the split string ('#')}

\item{pos}{which piece to keep (2)}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{   a character vector
}\if{html}{\out{</div>}}
}
\description{
extract unique barcodes from cellColData in an archR project
}
\examples{

inDEM <- archRbarCodes(MLLDEM)
inK27 <- archRbarCodes(MLLK27)
table(inDEM \%in\% inK27)
table(inK27 \%in\% inDEM) 

}
