% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archRgeneScoreMatrix.R
\name{archRgeneScoreMatrix}
\alias{archRgeneScoreMatrix}
\title{reformat ArchR GeneScoreMatrix rowdata into a GRanges}
\usage{
archRgeneScoreMatrix(proj, genome = NULL)
}
\arguments{
\item{proj}{the ArchRproject holding the GeneScoreMatrix}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{   the gene score data as a RangedSummarizedExperiment
}\if{html}{\out{</div>}}
}
\description{
reformat ArchR GeneScoreMatrix rowdata into a GRanges
}
