% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archRgeneScoresWithRanges.R
\name{archRgeneScoresWithRanges}
\alias{archRgeneScoresWithRanges}
\title{reformat ArchR GeneScoreMatrix rowdata into a GRanges}
\usage{
archRgeneScoresWithRanges(se, genome = NULL)
}
\arguments{
\item{se}{the SummarizedExperiment holding the GeneScoreMatrix}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{   the same data as a RangedSummarizedExperiment
}\if{html}{\out{</div>}}
}
\description{
reformat ArchR GeneScoreMatrix rowdata into a GRanges
}
