% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archRmetadata.R
\name{archRmetadata}
\alias{archRmetadata}
\title{compile ArchR metadata to pack into an SCE}
\usage{
archRmetadata(proj, lst = NULL)
}
\arguments{
\item{proj}{an ArchRproject}

\item{lst}{slots to mine for metadata (see details)}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{         a list of metadata items
}\if{html}{\out{</div>}}
}
\description{
compile ArchR metadata to pack into an SCE
}
\details{
\if{html}{\out{<div class="sourceCode">}}\preformatted{        defaults to mining 'projectMetadata',
                 'sampleMetadata', and 'cellMetadata'
}\if{html}{\out{</div>}}
}
