% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archRpeakMatrix.R
\name{archRpeakMatrix}
\alias{archRpeakMatrix}
\title{add ArchR peakSet ranges to a PeakMatrix-exported SE}
\usage{
archRpeakMatrix(proj, genome = NULL)
}
\arguments{
\item{proj}{the ArchR project, which MUST HAVE A peakSet (!)}

\item{genome}{the genome to annotate the peakSet from (NULL)}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{     a RangedSummarizedExperiment-housed PeakMatrix, or NULL
}\if{html}{\out{</div>}}
}
\description{
add ArchR peakSet ranges to a PeakMatrix-exported SE
}
