% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archRtileSize.R
\name{archRtileSize}
\alias{archRtileSize}
\title{figure out what tile size was used for an ArchR iterative LSI run}
\usage{
archRtileSize(proj)
}
\arguments{
\item{proj}{the ArchRProject}
}
\value{
unique(h5read(ArrowFile, "TileMatrix/Info/Params")$tileSize)
}
\description{
figure out what tile size was used for an ArchR iterative LSI run
}
\details{
the above is an oversimplification (!), see code for grim details
}
