% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binarizeMat.R
\name{binarizeMat}
\alias{binarizeMat}
\title{pretty much what it sounds like: binarize a sparse matrix}
\usage{
binarizeMat(mat)
}
\arguments{
\item{mat}{the matrix}
}
\value{
a binary version of mat
}
\description{
pretty much what it sounds like: binarize a sparse matrix
}
