% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkProjection.R
\name{checkProjection}
\alias{checkProjection}
\title{given a matrix of data and its LSI representation, check cor(projected, LSI)}
\usage{
checkProjection(mat, LSI = NULL, samp = 1000, MSE = FALSE)
}
\arguments{
\item{mat}{binarized tilematrix from an ArchR project, a project, or an SCE}

\item{LSI}{if 'mat' is not an ArchR project, a complete ArchR LSI object}

\item{samp}{how many cells to sample for the comparison? (1000)}

\item{MSE}{return mean squared error instead of correlations? (FALSE)}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{ column-wise correlations or diffs of projected data vs. matSVD
}\if{html}{\out{</div>}}
}
\description{
given a matrix of data and its LSI representation, check cor(projected, LSI)
}
\details{
\if{html}{\out{<div class="sourceCode">}}\preformatted{merge this into archRiSEE 
}\if{html}{\out{</div>}}
}
