% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downsample.R
\name{downsample}
\alias{downsample}
\title{block-randomized sampling to ensure somewhat representative comparisons}
\usage{
downsample(x, g, maxN = Inf, INDEX = 2, quietly = FALSE, ...)
}
\arguments{
\item{x}{a rectangular object that inherits from SummarizedExperiment}

\item{g}{grouping factor for each row/column, or its name in colData}

\item{maxN}{maximum size per-subset? (Inf; use the smallest subset size)}

\item{INDEX}{block-downsample on columns (2) or rows (1)? (2)}

\item{quietly}{suppress warnings about unsampled metadata? (FALSE)}

\item{...}{additional parameters to pass to sample(), e.g. replace=TRUE}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{   a subset of x, downsampled to match the smallest N(g)
}\if{html}{\out{</div>}}
}
\description{
block-randomized sampling to ensure somewhat representative comparisons
}
\details{
\if{html}{\out{<div class="sourceCode">}}\preformatted{  use set.seed beforehand if you want reproducibility.
}\if{html}{\out{</div>}}
}
\examples{

ncells <- 100
u <- matrix(rpois(20000, 5), ncol=ncells)
v <- log2(u + 1)
library(SummarizedExperiment)
se <- SummarizedExperiment(assays=list(counts=u, logcounts=v))
rownames(se) <- paste0("feature", seq_len(nrow(se)))
colnames(se) <- paste0("subject", seq_len(ncol(se)))
clusts <- factor(1:5) 
se$Clusters <- sample(clusts, ncol(se), replace=TRUE)
show(se) 

# without replacement:
set.seed(1234)
(downsample(se, "Clusters"))

# with replacement:
set.seed(1234)
(downsample(se, "Clusters", replace=TRUE))

}
