% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logTfIdf.R
\name{logTfIdf}
\alias{logTfIdf}
\title{ArchR-style log1p(TF-IDF) transformation but without mandatory binarization}
\usage{
logTfIdf(mat, prune = 1, idf = NULL, binarize = TRUE)
}
\arguments{
\item{mat}{a (sparse) matrix of counts (cells as columns) or an SE}

\item{prune}{prune (cells >= prune)? (1, by default)}

\item{idf}{a pre-trained idf or logTfIdf result (NULL; compute idf)}

\item{binarize}{binarize the values for TF-IDF? (TRUE)}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{     a log1p(TF-IDF) version of the (counts) matrix (see Details)
}\if{html}{\out{</div>}}
}
\description{
ArchR-style log1p(TF-IDF) transformation but without mandatory binarization
}
\details{
\if{html}{\out{<div class="sourceCode">}}\preformatted{    TF == (term / sum(terms_in_document));
             IDF == (1 / sum(documents_with_term_not_zero));
             currently only assays(mat)[1] gets transformed for SE/SCEs;
             provided or generated model idf placed in attr(mat, 'idf')
}\if{html}{\out{</div>}}
}
\seealso{
\if{html}{\out{<div class="sourceCode">}}\preformatted{    text2vec::TfIdf
}\if{html}{\out{</div>}}
}
