% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projectAssays.R
\name{projectAssays}
\alias{projectAssays}
\title{project a (subset of an) SCE into a space defined by another}
\usage{
projectAssays(x, y, z = "UMAP_projected")
}
\arguments{
\item{x}{SCE to be projected}

\item{y}{SCE with projection features and saved UMAP model}

\item{z}{reducedDim name to store the result in object x}
}
\value{
x, with a new reducedDim
}
\description{
project a (subset of an) SCE into a space defined by another
}
\details{
it's more fiddly than that, usually.  See examples.
}
