% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projectLSI.R
\name{projectLSI}
\alias{projectLSI}
\title{stolen directly from ArchR, with simplifications for the default settings}
\usage{
projectLSI(mat, LSI)
}
\arguments{
\item{mat}{a sparse Matrix, M rows x N columns}

\item{LSI}{an ArchR LSI object with nDimensions = K}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{     an N x K projection matrix 
}\if{html}{\out{</div>}}
}
\description{
stolen directly from ArchR, with simplifications for the default settings
}
