% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparsity.R
\name{sparsity}
\alias{sparsity}
\title{compute sparsity for a Matrix, SummarizedExperiment, or SingleCellExperiment}
\arguments{
\item{object}{the object whose sparsity we would like to evaluate}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{   1 - (length(object@x)/prod(dim(object))), i.e. (1 - density)
}\if{html}{\out{</div>}}
}
\description{
compute sparsity for a Matrix, SummarizedExperiment, or SingleCellExperiment
}
\details{
\if{html}{\out{<div class="sourceCode">}}\preformatted{  just adding a method for Matrix objects to RcppML's generic.
           if a SummarizedExperiment is passed, the first assay is used.
}\if{html}{\out{</div>}}
}
