% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stack.R
\name{stack}
\alias{stack}
\title{stack (rbind) two SCE-like objects for the same cells to integrate}
\usage{
stack(x, y, NMF = TRUE, UMAP = TRUE, k = 30, ...)
}
\arguments{
\item{x}{the first object (its rows will be on top)}

\item{y}{the second object (its rows will be below y)}

\item{NMF}{run NMF on the merged feature set? (TRUE)}

\item{UMAP}{run UMAP on the NMF results? (TRUE; see details)}

\item{k}{rank of the NMF decomposition to shoot for (default: 30)}

\item{...}{currently unused parameters for convenience functions}

\item{an}{object of the same type as x and y, if all goes well}
}
\description{
stack (rbind) two SCE-like objects for the same cells to integrate
}
\details{
\if{html}{\out{<div class="sourceCode">}}\preformatted{rbind()'ing objects is easy. Using the result isn't always so
         easy. We assume that people want to use NMF and UMAP to get a
         somewhat interpretable look at the combined data, but these are
         just two of many, many possible ways to approach the task. Most
         of the time, users will want to run archRtoSCE on a couple of 
         archR projects with how='feats' and feats=commonFeaturesGRanges
         to ensure that the desired features from both modalities are 
         present. After that the process tends to be iterative, which is
         the whole point of using iSEE to interact with the results.
}\if{html}{\out{</div>}}
}
