% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stackAssays.R
\name{stackAssays}
\alias{stackAssays}
\title{stack (rbind) multiple SCE-like objects for the same cells to integrate}
\usage{
stackAssays(lst, ...)
}
\arguments{
\item{lst}{a list of SCE-like objects with identical colnames()}

\item{...}{currently unused parameters for convenience functions}

\item{an}{object of the same type as x and y, if all goes well}
}
\description{
stack (rbind) multiple SCE-like objects for the same cells to integrate
}
\details{
\if{html}{\out{<div class="sourceCode">}}\preformatted{rbind()'ing objects is easy. Using the result isn't always so
         easy. We assume that people want to use NMF and UMAP to get a
         somewhat interpretable look at the combined data, but these are
         just two of many, many possible ways to approach the task. Most
         of the time, users will want to run archRtoSCE on a couple of 
         archR projects with how='feats' and feats=commonFeaturesGRanges
         to ensure that the desired features from both modalities are 
         present. After that the process tends to be iterative, which is
         the whole point of using iSEE to interact with the results.
         Currently, NMF and UMAP on the combined features is disabled
         because the default binary output makes no sense with NMF.
         If you plan to use raw or lognormalized FragmentCounts again, 
         make sure that the object you want to use these on is lst[[1]],
         else you'll have to swapAltExp(stacked, whatever) later on.
}\if{html}{\out{</div>}}
}
