% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizeOver.R
\name{summarizeOver}
\alias{summarizeOver}
\title{given an SCE and a set of features, summarize counts over those features}
\usage{
summarizeOver(x, gr, asy = NULL, BPPARAM = MulticoreParam())
}
\arguments{
\item{x}{a SummarizedExperiment, SingleCellExperiment, or similar}

\item{gr}{a GRanges of features, or something coercible to GRanges}

\item{asy}{which assay to summarize? (whichever one is first)}

\item{BPPARAM}{BiocParallel parameters (default is MulticoreParam)}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{   a SummarizedExperiment holding the summarized counts
}\if{html}{\out{</div>}}
}
\description{
given an SCE and a set of features, summarize counts over those features
}
\details{
\if{html}{\out{<div class="sourceCode">}}\preformatted{  the result is an SE because you'll want to make it an altExp.
           you'll probably also want to run logNormCounts on the result.
}\if{html}{\out{</div>}}
}
\examples{

\dontrun{
  library(rtracklayer) 
  K27feats <- import("NBMK27_50kbp_LSIfeats.hg38.bed")
  (MLLK27_50kb <- summarizeOver(MLLK27, K27feats))
  MLLK27_50kb <- logNormCounts(MLLK27_50kb)
  altExp(MLLK27, "K27_50kb_forNBM") <- MLLK27_50kb
}

}
\seealso{
\if{html}{\out{<div class="sourceCode">}}\preformatted{  scuttle::logNormCounts
}\if{html}{\out{</div>}}

\if{html}{\out{<div class="sourceCode">}}\preformatted{  SSBtools
}\if{html}{\out{</div>}}
}
