% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sweepSparse.R
\name{sweepSparse}
\alias{sweepSparse}
\title{fast row/column sweep for sparse matrices}
\usage{
sweepSparse(x, MARGIN, STATS, FUN = "/")
}
\arguments{
\item{x}{a (sparse) matrix}

\item{MARGIN}{margin to be swept}

\item{STATS}{what to sweep out}

\item{FUN}{how to sweep it out ('/')}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{     x, with stats swept out (divided through by default)
}\if{html}{\out{</div>}}
}
\description{
fast row/column sweep for sparse matrices
}
