% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unitScale.R
\name{unitScale}
\alias{unitScale}
\alias{rowUnitScale}
\title{scale factors (e.g. NMF factors) to unit distance}
\usage{
unitScale(x, MARGIN = 1)

rowUnitScale(x, ...)
}
\arguments{
\item{x}{a (sparse) matrix}

\item{MARGIN}{1: rows, 2: columns (1; unnecessary for row/colUnitScale())}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{     x, scaled 0-1
}\if{html}{\out{</div>}}
}
\description{
scale factors (e.g. NMF factors) to unit distance
}
