% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unitScale.R, R/zScore.R
\name{colUnitScale}
\alias{colUnitScale}
\alias{zScore}
\alias{rowZscores}
\alias{colZscores}
\title{scale factors (e.g. PCA) to Z-scores}
\usage{
colUnitScale(x, ...)

zScore(x, MARGIN = 1)

rowZscores(x, ...)

colZscores(x, ...)
}
\arguments{
\item{x}{a (sparse) matrix}

\item{MARGIN}{1: rows, 2: columns (1; unnecessary for row/colZscores())}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{     x, Z-scored
}\if{html}{\out{</div>}}
}
\description{
scale factors (e.g. PCA) to Z-scores
}
