% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addIgvTrack.R
\name{addIgvTrack}
\alias{addIgvTrack}
\title{wrapper to add a weights track to an igvR instance}
\usage{
addIgvTrack(x, trk, igv, co = "violet", a = TRUE, ...)
}
\arguments{
\item{x}{SingleCellExperiment with weights in mcols(rowRanges)$trk}

\item{trk}{name of the column in mcols(rowRanges(x)) with weights to plot}

\item{igv}{igvR object with the appropriate genome to match x}

\item{co}{color for the track (violet)}

\item{a}{autoscale the track? (TRUE)}

\item{...}{additional parameters passed to igvR::displayTrack}
}
\value{
none, called for side effects
}
\description{
wrapper to add a weights track to an igvR instance
}
\examples{
library(igvR)
igv <- igvR()
mcols(SCE)$aTrack <- rpois(nrow(SCE), lambda=50)
addIgvTrack(SCE, "aTrack", igv) 

}
\seealso{
igvR::GRangesQuantitativeTrack

igvR::displayTrack
}
