% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archRgenome.R
\name{archRgenome}
\alias{archRgenome}
\title{figure out which genome was used by ArchR}
\usage{
archRgenome(proj)
}
\arguments{
\item{proj}{an ArchRproject}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{         a character string indicating the genome
}\if{html}{\out{</div>}}
}
\description{
figure out which genome was used by ArchR
}
