% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archRiterLSI.R
\name{archRiterLSI}
\alias{archRiterLSI}
\title{retrieve relevant IterativeLSI bits from an ArchR project}
\usage{
archRiterLSI(proj, LSIdim = "IterativeLSI", keep = NULL)
}
\arguments{
\item{proj}{the ArchRProject}

\item{LSIdim}{name of the IterativeLSI reducedDims entry ("IterativeLSI")}

\item{keep}{these ("matSVD","outliers","LSIFeatures","u","d","v","idf")}
}
\value{
a list with embeddings, outliers to exclude, features GRanges, etc.
}
\description{
retrieve relevant IterativeLSI bits from an ArchR project
}
