% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archRtoSCE.R
\name{archRtoSCE}
\alias{archRtoSCE}
\title{turn an ArchR project into a SingleCellExperiment}
\usage{
archRtoSCE(proj, tile = 500, addNMF = FALSE, k = 30, ...)
}
\arguments{
\item{proj}{an ArchRproject}

\item{tile}{tile size (500)}

\item{addNMF}{add an NMF decomposition of logCounts? (FALSE)}

\item{k}{rank for the NMF decomposition on logCounts (30)}

\item{...}{additional arguments to pass to ArchR::addTileMatrix}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{         a SingleCellExperiment 
}\if{html}{\out{</div>}}
}
\description{
The default tile size is 500bp, just like ArchR's default.
}
\details{
presumes LSI and UMAP. you will almost certainly want to tweak this.
}
