% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iSEEarchR.R
\name{iSEEarchR}
\alias{iSEEarchR}
\title{convenience wrapper for iSEE on an ArchR-derived singlecellexperiment}
\usage{
iSEEarchR(x, colorColumn = "Clusters")
}
\arguments{
\item{x}{a SingleCellExperiment with UMAP and LSI reducedDims}

\item{colorColumn}{the colData column to color plots ("mitoCluster")}
}
\description{
Note that this is a quick and dirty affair so don't expect much.
}
\details{
presumes NMF and UMAP. you will almost certainly want to tweak this.
}
