% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sweepSparse.R
\name{sweepSparse}
\alias{sweepSparse}
\title{fast column sweep for sparse matrices}
\usage{
sweepSparse(mat, stats, fun = "/")
}
\arguments{
\item{mat}{a (sparse) matrix}

\item{stats}{what to sweep out}

\item{fun}{how to sweep it out ('/')}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{     mat, with colSums swept out (divided through, by default)
}\if{html}{\out{</div>}}
}
\description{
fast column sweep for sparse matrices
}
