% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variantFreq.R
\name{variantFreq}
\alias{variantFreq}
\title{tabulate (usually) offtarget vs. ontarget variants}
\usage{
variantFreq(
  SE,
  grouping = NULL,
  loci = NULL,
  altexp = "mtVariants",
  assay = "allele_frequency",
  cutoff = 0.05
)
}
\arguments{
\item{SE}{SingleCellExperiment or a SummarizedExperiment (see Details)}

\item{grouping}{names of factors in colData(SE) to group by (or NULL)}

\item{loci}{rownames to include (or NULL)}

\item{altexp}{'mtVariants', by default; ignored unless SE is an SCE}

\item{cutoff}{0.05 (five percent), by default; if 0, just return AFs}

\item{asy}{'allele_frequency', by default; can also be NULL (assay 1)}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{     a vector or table of cells passing the cutoff at each locus
}\if{html}{\out{</div>}}
}
\description{
tabulate (usually) offtarget vs. ontarget variants
}
\details{
\if{html}{\out{<div class="sourceCode">}}\preformatted{    this function is not very clever, it just sets up HET calls
             for tabulating by (eg) cellLine and edit status
}\if{html}{\out{</div>}}
}
\examples{

\dontrun{

  # this takes time, even though it IS subsetted
  system.time(ewsSub <- readRDS("ewsSub.rds"))
  #    user  system elapsed 
  # 200.518   5.787 206.487 
  
  grouping <- c("cellLine", "edited")
  loci <- c("11922G>A", "11925C>T", "11946C>T")
  res <- variantFreq(ewsSub, grouping=grouping, loci=loci)
  names(res) 
  # [1] "11922G>A" "11925C>T" "11946C>T" "cellLine" "edited"  
  
  (names(res)[1:3] <- paste0("chrM_", sub('>', '_', names(res)[1:3])))
  # [1] "chrM_11922G_A" "chrM_11925C_T" "chrM_11946C_T"
  
  with(res, table(chrM_11922G_A, edited, cellLine))  # on-target
  with(res, table(chrM_11925C_T, edited, cellLine))  # off-target 

}


}
\seealso{
archRtoSCE
}
